fix-attn

+++ zeromodels/models/glm4v/glm4v_vision_layers.py
@@ -276,7 +276,7 @@ class Glm4vVisionAttention(layers.Layer):
 
         attn = ops.matmul(query, ops.transpose(key, (0, 1, 3, 2))) * self.scaling
         if attention_mask is not None:
-            attn = attn + attention_mask
+            attn = attn + ops.cast(attention_mask, attn.dtype)
         attn = ops.cast(ops.softmax(ops.cast(attn, "float32"), axis=-1), query.dtype)
         out = ops.matmul(attn, value)
         out = ops.transpose(out[0], (1, 0, 2))
